Change spell slots to use a dedicated array and custom menu, rather than try to rely on Skyrim's favouriting system.
Add option for number of slots to scale with magic skill.

If slots are used, give player a lesser power that can be used in inns/homes/etc. Power brings up a menu which allows assigning spells to slots.

Check for equipped spells in OnUpdate, since unfortunately OnObjectEquipped seems not to work with spells.